The core and the persistence build without Qt
419 QString to std::string, and with them the last of the Qt in
src/core and src/persistence: QFile and QTextStream to ifstream,
QByteArray to string, QElapsedTimer to chrono, QHash to map,
QRegularExpression to regex, qDebug and qWarning to cout and cerr,
Q_UNUSED to a cast to void, qAcos and qMin to their std equivalents.

The API was small: 141 QString::number went to the primitive from the
earlier commit, 35 toStdString simply disappeared, and there were only
eight arg() to write out by hand. text::join is new, because QStringList
had one and the alternative is a loop with a first-time flag at every call.

Two things fell out on the way. ExpressionTree::es_letra built a regular
expression and a one-character string to ask whether a character is a
letter; it is two range comparisons now, and unlike isalpha they do not
depend on the locale, which under another one could start accepting
accented letters the lexer has no rule for. And the reader's realVector
was a second implementation of text::reals, whole-token validation
included.

The acceptance test passes: qftbx_core_noqt compiles both directories
with the Qt include paths taken away, and Qt6::Core is off the real
target. Zero Qt includes remain; the QString and QMap a grep still finds
are inside comments, describing what things used to be. Qt stays linked to
the test binaries as scaffolding only - QTemporaryDir for the files two
tests write - and the proof of the core's independence is the noqt target,
not the tests' link line.

A regression came with it, and closing it was the interesting part. The
heavy tests went from 1631/1658/1693 ms to 1975/1975/2305, twenty percent
slower. Two causes, both measured rather than guessed.

evaluateCached built its key - a pair of a string and a vector of strings
- on every single evaluation in order to look it up. QString made that a
refcount bump; std::string makes it a real copy of the expression and
every name, and a template sweep evaluates the plant thousands of times
per frequency. There is now a one-entry memo in front of the map that
compares POINTERS, since a hit requires the caller to hand over the same
objects, which is exactly what a Parameter or a FreeForm evaluating itself
does.

The second was mine: removing QFile, I read files with
istreambuf_iterator, which is character by character through a virtual
call. Callgrind put 5.8% of a template run in _M_get, _S_is_eof and
eq_int_type - to read one file. It is sized and read in one go now, and
the median is back in the old band.

The lesson is the one foreach taught two commits ago. Qt's copy-on-write
was not only hiding a dangling pointer; it was hiding copies. Taking it
away means looking at everywhere something used to be free.

+++ src/gui/boundary_union_viewer.cpp
@@ -277,7 +277,7 @@ void BoundaryUnionViewer::addFrequencyRow(QColor color, qint32 pos){
     QMetaObject::connectSlotsByName(widget);
 
 
-    if (pos < omega->size()){
+    if (pos < static_cast<std::int32_t>(omega->size())){
         checkBox->setText(QString::number(omega->at(pos)));
     } else {
         checkBox->setText(tr("union"));
